+++ NEWS.md
@@ -8,6 +8,7 @@
 
 ## New features
 
+- `vd_a()` for Vargha and Delaney's *A* dominance effect size.
 - `cohens_u1()`, `cohens_u2()`, `d_to_u1()`, and `d_to_u2()` added for Cohen's U1 and U2
 - `mahalanobis_d()` for multivariate standardized differences.